Final pages to /app
