+++ app/alpha/create/[[...postIdArr]].tsx
