+++ app/alpha/layout.tsx
@@ -1,0 +1,11 @@
+export const metadata = {
+  title: "🚨 WIP - Things will break",
+  robots: {
+    follow: false,
+    index: false,
+  },
+};
+
+export default function Alpha({ children }: { children: ChildNode }) {
+  return <>{children}</>;
+}
